+++ src/main/scala/rift2Core/L1Cache/ProbeUnit.scala
@@ -17,7 +17,7 @@ class Info_probe_req extends Bundle {
 /**
   * ProbeUnit will accept probe request from l2cache and forward it to l1cache to resp data
   */
-class ProbeUnit(edge: TLEdgeOut)(implicit p: Parameters) extends L1CacheModule {
+class ProbeUnit(edge: TLEdgeOut, id: Int)(implicit p: Parameters) extends L1CacheModule {
   val io = IO(new Bundle {
     val cache_probe = Flipped(new DecoupledIO(new TLBundleB(edge.bundle)))
     val req = new DecoupledIO(new Info_probe_req)
@@ -33,6 +33,7 @@ class ProbeUnit(edge: TLEdgeOut)(implicit p: Parameters) extends L1CacheModule {
   probe_fifo.io.enq.bits := io.cache_probe.bits.address
   io.cache_probe.ready := probe_fifo.io.enq.ready
 
+  // assert( ~(io.cache_probe.fire & io.cache_probe.bits.source =/= id.U), "Assert Failed at probe-unit, source id mis-match" )
 
   io.req.valid := probe_fifo.io.deq.valid
   probe_fifo.io.deq.ready := io.req.ready
