+++ src/main/scala/rift2Core/L1Cache/cache_tag.scala
@@ -23,10 +23,12 @@ import chisel3.util._
 import base._
 
 
-class Cache_tag( dw: Int, aw: Int, bk: Int, cb: Int, cl: Int, nm: Int ) {
-  val addr_lsb = log2Ceil(dw*bk/8)
+class Cache_tag( dw: Int, aw: Int, cb: Int, cl: Int, bk: Int ) {
+  val addr_lsb = log2Ceil(dw/8)
   val line_w   = log2Ceil(cl)
-  val tag_w   = aw - addr_lsb - line_w - log2Ceil(nm)
+  val bk_w     = log2Ceil(bk)
+  val tag_w   = aw - addr_lsb - line_w - bk_w
+  
 
 
   val tag_addr_r = Wire(UInt(aw.W))
@@ -37,8 +39,8 @@ class Cache_tag( dw: Int, aw: Int, bk: Int, cb: Int, cl: Int, nm: Int ) {
   val tag_info_r = Wire( Vec(cb, UInt(tag_w.W)) )
 
 
-  val addr_sel_w = tag_addr_w(addr_lsb+line_w-1, addr_lsb)
-  val addr_sel_r = tag_addr_r(addr_lsb+line_w-1, addr_lsb)
+  val addr_sel_w = tag_addr_w(addr_lsb+bk_w+line_w-1, addr_lsb+bk_w)
+  val addr_sel_r = tag_addr_r(addr_lsb+bk_w+line_w-1, addr_lsb+bk_w)
 
   val tag_info_w = tag_addr_w(aw-1, aw-tag_w)
