+++ src/main/scala/rift2Core/backend/Alu.scala
@@ -38,46 +38,42 @@ class Alu extends Module {
   io.alu_exe_iwb <> alu_exe_iwb_fifo.io.deq
   alu_exe_iwb_fifo.reset := reset.asBool | io.flush
 
-  def iss_ack = io.alu_iss_exe.valid
 
+  val is_32w = io.alu_iss_exe.bits.param.is_32w
+  val is_usi = io.alu_iss_exe.bits.param.is_usi
 
-  def is_32w = io.alu_iss_exe.bits.param.is_32w
-  def is_usi = io.alu_iss_exe.bits.param.is_usi
-
-  def op1 = io.alu_iss_exe.bits.param.dat.op1
-  def op2 = io.alu_iss_exe.bits.param.dat.op2
+  val op1 = io.alu_iss_exe.bits.param.dat.op1
+  val op2 = io.alu_iss_exe.bits.param.dat.op2
 
   def cutTo32(in: UInt) = Cat(Fill( 32, in(31) ), in(31, 0))
 
-  // def op1_32w = Cat( Fill(32,op1(31)), op1(31,0))
-  // def op2_32w = Cat( Fill(32,op2(31)), op2(31,0))
 
-  def adder_res = op1 + op2
-  def adder_res_32w = cutTo32(adder_res)
-  def alu_add_res = Mux(is_32w, adder_res_32w, adder_res)
+  val adder_res = op1 + op2
+  val adder_res_32w = cutTo32(adder_res)
+  val alu_add_res = Mux(is_32w, adder_res_32w, adder_res)
 
-  def slt_sign_res = Mux( op1.asSInt < op2.asSInt, 1.U, 0.U )
-  def slt_unsi_res = Mux( op1 < op2, 1.U, 0.U )
+  val slt_sign_res = Mux( op1.asSInt < op2.asSInt, 1.U, 0.U )
+  val slt_unsi_res = Mux( op1 < op2, 1.U, 0.U )
 
-  def alu_slt_res = Mux(is_usi, slt_unsi_res, slt_sign_res)
+  val alu_slt_res = Mux(is_usi, slt_unsi_res, slt_sign_res)
 
-  def alu_xor_res = op1 ^ op2;
-  def alu_or_res  = op1 | op2;
-  def alu_and_res = op1 & op2;
+  val alu_xor_res = op1 ^ op2;
+  val alu_or_res  = op1 | op2;
+  val alu_and_res = op1 & op2;
 
 
 
-  def shift_op2 = Mux(is_32w, op2(4,0), op2(5,0))
+  val shift_op2 = Mux(is_32w, op2(4,0), op2(5,0))
 
-  def alu_sll_res  = Mux( is_32w, cutTo32( op1 << shift_op2), op1 << shift_op2 )
+  val alu_sll_res  = Mux( is_32w, cutTo32( op1 << shift_op2), op1 << shift_op2 )
 
-  def alu_srl_res  = Mux( is_32w, cutTo32(op1(31,0) >> shift_op2), op1 >> shift_op2)
+  val alu_srl_res  = Mux( is_32w, cutTo32(op1(31,0) >> shift_op2), op1 >> shift_op2)
 
-  def sra_op1_128w = Mux( is_32w, Cat( Fill(96, op1(31)), op1(31,0) ), Cat( Fill(64, op1(63)), op1(63,0) ) )
-  def alu_sra_res  = Mux( is_32w, cutTo32(sra_op1_128w >> shift_op2), sra_op1_128w >> shift_op2 )
+  val sra_op1_128w = Mux( is_32w, Cat( Fill(96, op1(31)), op1(31,0) ), Cat( Fill(64, op1(63)), op1(63,0) ) )
+  val alu_sra_res  = Mux( is_32w, cutTo32(sra_op1_128w >> shift_op2), sra_op1_128w >> shift_op2 )
 
 
-  def res = MuxCase(DontCare, Array(
+  val res = MuxCase(DontCare, Array(
     io.alu_iss_exe.bits.fun.add -> alu_add_res,
     io.alu_iss_exe.bits.fun.slt -> alu_slt_res,
     io.alu_iss_exe.bits.fun.xor -> alu_xor_res,
