--- src/main/scala/rift2Core/privilege/csrFiles/CsrFiles_D.scala.bk
@@ -1,51 +1,0 @@
-
-/*
-  Copyright (c) 2020 - 2022 Wuhan University of Technology <295054118@whut.edu.cn>
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package rift2Core.privilege.csrFiles
-
-import chisel3._
-import chisel3.util._
-import rift2Core.define._
-import base._
-
-// trait CsrFiles_D {  this: BaseCommit =>
-//   val ResetReq  = Wire(Bool())
-
-
-//   val is_debug_interrupt = Wire(Bool())
-
-//   val is_ebreak_retired = Wire(Bool())
-//   val is_single_step = Wire(Bool())
-//   val is_trigger = Wire(Bool())
-//   val is_halt_int = Wire(Bool())
-
-  
-
-
-
-
-
-
-
-
-
-
-
-// }
-
-
-
